Lighting and collision implemented (#38)
* collision implemented and tested

* Normal coordinates are now used to generate lighting. Lighting intensty and direction still needs some tweaking

+++ ProftaakPeriode4/ProftaakPeriode4/LaneGeneratorComponent.cpp
@@ -61,9 +61,9 @@ Mesh* Lane::RandomMesh()
 	return _meshes.at(index);
 }
 
-int Lane::getWidth()
+float Lane::getWidth()
 {
-	int largest = 1;
+	float largest = 1;
 	for (int i = 0; i < _queue.size(); i++)
 		if (_queue[i]->_width > largest)
 			largest = _queue[i]->_width;
@@ -85,7 +85,7 @@ LaneGeneratorComponent::~LaneGeneratorComponent()
 
 void LaneGeneratorComponent::Draw()
 {
-	int alreadyDrawn = 0;
+	float alreadyDrawn = 0;
 	for (int i = 0; i < _lanes.size(); i++) {
 		_lanes.at(i)->Draw(alreadyDrawn + (i*_spaceBetween));
 		alreadyDrawn += _lanes[i]->getWidth();
