+++ ProftaakPeriode4/ProftaakPeriode4/LaneGeneratorComponent.h
@@ -67,7 +67,7 @@ public:
 	 * todo test with multiple 3d models to be sure the alignment is correct
 	 * \return the Maximum width of the lane
 	 */
-	int getWidth();
+	float getWidth();
 
 	/**
 	 * The meshes that are randomly chosen (loaded in constructor)
